+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Replaced Calendar component's `hasHeader` property & `has-header` attribute with `hideHeader` & `hide-header` respectively.
 - Replaced Card component's `outlined` property with `elevated`.
 
+### Removed
+
+- Removed `igcOpening`, `igcOpened`, `igcClosing` and `igcClosed` events from Navigation drawer component.
+
 ## 1.0.0 - 2021-11-22
 
 Initial release of Ignite UI Web Components